Add three ProDOS API calls a2getmark() - MLI call $CF - Returns file position within an open file a2setmark() - MLI call $CE - Changes the file position within an open file a2seteof() - MLI call $D0 - Change the EOF position of an open file
Input to a2setmark() and a2seteof() require a pointer to a 64-bit DWORD implemented as a pair of sequential words.
Working with 24-bit values to use a2setmark() and a2seteof() is a pain when using a language that has a maximum integer value of $FFFF.  I've created some el cheapo 24-bit add and multiply routines to use in testing.
Next test will be to see if we really need to pass the 24-bit lengths as pointers or can we live with passing by value.

Did not create a version of the routines for apple 1 or apple 3 because I have no means to test them.